requirements.txt
python-dotenv~=1.1.0
openai~=1.84.0


